fix(#2454): move list snapshot to neutral agent_ops service
Extract list_response logic from api::handlers::query to
agent_ops::list_snapshot — a neutral typed in-process boundary
that both the API LIST wire handler and MCP instance queries call
directly, breaking the MCP→API compile dependency for this family.

Changes:
- agent_ops::list_snapshot: owns lock-drop-before-disk-I/O ordering,
  managed/external agent serialisation, dispatch_idle enrichment
- api::handlers::query::list_response: thin adapter calling
  agent_ops::list_snapshot (API LIST wire format preserved)
- api::mod.rs: remove dead list_response re-export (no callers)
- instance_queries.rs: list path uses agent_ops::list_snapshot;
  describe_instance accepts RuntimeContext, uses list_snapshot
  in-process (no api::call); runtime=None returns explicit error

Preserves: full/compact/evidence, external agents, metadata merge,
topic_id lookup, topic_binding_mode, operator_mode, is_self marker.


Agend-Agent: claude-aef7c0
Agend-Task: t-20260718154741427470-72215-31
Agend-Branch: fix/2454-runtime-instance-query-service
Agend-Issued-At: 2026-07-18T15:47:54.952553+00:00
